Seach latest or oldest 4chan posts

This function returns latest or oldest 4chan posts specified by searchterms

Usage

search_4chan_snippet(
  start_date = "",
  end_date = "",
  boards,
  text = "",
  subject = "",
  filename = "",
  ghost = "all",
  user_id = "",
  tripcode = "",
  type = "all",
  username = "",
  country = "",
  results = "all",
  show_only = "all",
  deleted = "all",
  capcode = "all",
  order = "asc",
  cool = 0,
  page = 1,
  result_type = "snippet"
)

Arguments

start_date

String, Start date of post search. Format: "YYY-MM-DD", Default: ''

end_date

String, End date of post search. Format: "YYY-MM-DD", Default: ''

boards

String or string vector of the 4chan board.
Available boards are: "adv", "plebs", "hr", "tg", "tv", "x", "s4s", "pol", "o", "trv", "f", "sp", "mlpol", "mo".

text

String, Searchterms in post text. Regex possible, Default: ''

subject

String, Searchterms in post subject. Regex possible, Default: ''

filename

String, Searchterms in filename. Regex possible, Default: ''

ghost

String, Filter by: "all", "only", "none" , Default: 'all'

user_id

Integer, Filter by user_id, Default: ''

tripcode

String, Filter by tripcode, Default: ''

type

String, Filter by post type: "all", "sticky", "op", "posts", Default: 'all'

username

String, Search username, Default: ''

country

String, Filter by country. Format: ISO 3166-1 Alpha-2 country code, Default: ''

results

String, Return post matches or the thread they are found in. Possible values: "all", "thread", Default: ''

show_only

String, Show only posts containing data types: "all", "text", "image", Default: 'all'

deleted

String, Filter deletion status: "all", "not-deleted", "deleted", Default: 'all'

capcode

String, Filter author type: "all", "user", "ver", "mod", "dev", "admin", "manager", "founder", Default: 'all'

order

String, Return posts ordered by latest or oldest: "asc", "desc", Default: 'asc'

cool

Integer (seconds), The 4plebs API includes an API rate limit of 5 requests/min for the archive search. For multiple searches a cool-down is recommended , Default: 0

page

Index of search result, Default: 1

result_type

String, Return output on the post level or only the number of posts found: "snippet", "results_num", Default: 'snippet'

Value

Dataframe with details on the most recent/oldest posts found.

Details

Variables in API output:

thread_id: 4chan ID of the thread the post is situated in
doc_id: 4chan document ID
num: 4chan post ID
subnum: Binary, 1: Ghost Post, 0: Non-Ghost Post
op: Binary, 1: Opening Post, 0: Reply Post
timestamp: Time sent in Posix time
fourchan_date: Time sent
name: Author name
title: Post title
referencing_comment: The post_id this post is quoting
comments: Text of the post
poster_country: Author country
nreplies: Number of replies formatted: Boolean, Has this post been formatted?
media_link: Download link to the media (e.g. images) that have been shared in the post

If result_type is "results_num":
total_found: All number of total posts found with this query
actual_query_result: The 4plebs API limits the search results shown. This limit is 100,000 results. If the number of total posts found exceeds this limit, the actual_query_results are capped.

See also

Examples

if (FALSE) {
t <- search_4chan_snippet(boards = "adv", text = "kitties",
show_only = "image")

search_4chan_snippet(boards = "pol", country = "BR",
result_type = "results_num")
}